WIP: remove NIH_BEGIN/END_EXTERN from the headers.
NIH_BEGIN/END_EXTERN expands to 'extern "C"' macro.
This is only relevant if these header files are going to
get included from the CPP file, which I don't see it
happening in the near future.

For the sake of convenience, I will remove these
extern definition from the headers instead of
redefining the equivalent on ureadahead side.