replace nih_assert with normal assert
This is part of ongoing effort to decouple libnih from the ureadahead.

nih_assert performs condition check, print out the fatal log
if it fails and then immediately calls abort().

This behavior is almost the same as assert defined by C standard
except the formatting of the message. this makes C standard assert
a viable alternative.

nih_assert_not_reached is replaced with __builtin_unreachable.

Signed-off-by: Soichiro Fujii <soichirof@google.com>