Fix: gracefully handle pack size ABI incompatibility
While reading a pack file, ureadahead uses `sizeof num_groups` and
and `sizeof num_paths` when reading these members.
These members have a `size_t` type, which can be 4 bytes or 8 bytes
depending on the word size of the running environment. If the word
size changes (e.g. due to OS upgrade) and pack file persists
in-between the upgrade, it can read invalid value which can lead to
incorrect behavior, or crash.

Introduce the word size flag to the pack file, and properly validate
the expected word size for pack file.

Signed-off-by: Soichiro Fujii <soichirof@google.com>